coap/server/requirements.txt
setuptools
CoAPthon3>=1.0.1
elasticsearch>=7.0.5